requirements.txt
annotated-types==0.7.0
asttokens==2.4.1
backcall==0.2.0
blessed==1.20.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
comm==0.2.2
confection==0.1.4
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
en-core-web-sm==3.7.1
executing==2.0.1
faiss-gpu==1.7.2
filelock==3.13.1
fonttools==4.53.0
fsspec==2024.2.0
gpustat==1.1.1
huggingface-hub==0.23.0
idna==3.7
importlib-metadata==7.2.1
importlib-resources==6.4.0
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter-client==8.6.2
jupyter-core==5.7.2
kiwisolver==1.4.5
langcodes==3.4.0
language-data==1.2.0
logparser3==1.0.4
marisa-trie==1.1.1
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.550.52
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
packaging==24.0
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.2
preshed==3.0.9
prompt-toolkit==3.0.47
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic-core==2.18.2
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2022.3.2
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==2.7.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
thinc==8.2.3
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0+cu118
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.40.1
triton==2.3.0
typer==0.9.4
typing-extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
zipp==3.19.2
